chore(cronjob): move cronjob to jobs folder
